Handle missing credentials for Rubygems
When the user has not logged in to rubygems.org yet, the `gem push`
command asks for credentials interactively. Cheetah, which we use
to execute the command doesn't support this case, so for now we
just catch this case and advise the user to log in first by running
the command manually once.

As part of this fix the system tests where rearranged a bit to
minimize starting of containers, and some better error handling
was added.

Fixes #3