Execute commands via dedicated class in `assert`
`Assertion#assert` now takes an executor object which is used to
do the actual execution of commands. This way the dry run can be
implemented by passing a different executor object without the
code in the Assertion class being aware of it.

This also makes error handling more consistent. Errors in `assert`
are now signalled by throwing an `AssertionError` exception.